requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
pytest>=7.0.0
pillow>=9.0.0
# tkinter is built into Python, no additional installation needed 